add/change

+++ SoftUni/Python OOP - Exercises/exam_prep/Python OOP Exam - 16 Aug 2020/01/project/software/light_software.py
@@ -4,4 +4,4 @@ from project.software.software import Software
 class LightSoftware(Software):
 
     def __init__(self, name, capacity_consumption, memory_consumption):
-        super().__init__(name, 'Light', capacity_consumption * 1.5, memory_consumption * 0.5)
+        super().__init__(name, 'Light', int(capacity_consumption * 1.5), int(memory_consumption * 0.5))
